Aggiunta immagini dove mancavano

+++ Control/CAmministratore.php
@@ -246,12 +246,14 @@ class CAmministratore
 
             $usernameAmm = $session->readValue('username');
             if ($id!=-1){
-                $gruppo = $pm->load($id, 'FGruppo');
-                $campo = $gruppo->getCampo();
-                $partecipanti = $gruppo->getPartecipanti();
-                $amministratore = $pm->load($usernameAmm, 'FAmministratore');
-                $amministratore->rimborsaGruppo($partecipanti, $campo);
-                $pm->delete($id, 'FGruppo');
+                if($gruppo = $pm->load($id, 'FGruppo')) {
+                    ;
+                    $campo = $gruppo->getCampo();
+                    $partecipanti = $gruppo->getPartecipanti();
+                    $amministratore = $pm->load($usernameAmm, 'FAmministratore');
+                    $amministratore->rimborsaGruppo($partecipanti, $campo);
+                    $pm->delete($id, 'FGruppo');
+                }
             }
             header("Location: /PolisportivaDDD/Utente/home");
         }
